execution-plan: Use snapshot tests for program memory
This saves a "snapshot" of program memory (visualized as a table) into the repo. When you rerun the "draw cube via API" unit test, it'll compare the snapshot to what's on disk and fail if there's a change (like expectorate or twenty-twenty).

Use `cargo insta review` for reviewing changes and accepting/rejecting them.