Add the grackle variant of import
A new Instruction::ImportFiles is added in order to support importing
geometry files using KCL. The feature itself is split into two parts:

* Instruction::ImportFiles - Do the fs related machinery.
* Endpoint::ImportFiles - Do the network request.

The feature first began with an attempt at lower level filesystem-based
instructions, but it became quickly apparent KCVM isn't immediately
equipped to deal with doing an unknown amount of N instructions, in this case,
N file reads.

The upside is, the implementation becomes completely abstracted by the new
instruction.

The downside is, it's another instruction added to the instruction set.

SketchGroups have gone through the same course of development, so for now,
it's acceptable.