Remove Export::source_unit (#241)
This removes the 'source units', which were previously the user's choice of what 1 unit in the engine corresponded to in the real world. This is what is now fixed to 1mm, so a 1x1x1m cube would be represented as 1000x1000x1000 units in the engine. The user can select which unit they want the units in the exported file to be without changing the physical dimensions of the part in the [format configuration options](https://github.com/KittyCAD/modeling-api/blob/main/modeling-cmds/src/format/obj.rs#L59).

For example, if the user creates a part of length 1m, the format options choose whether that length is represented as 1 (m), 100 (cm), 1000 (mm), etc. in the exported file.