KCVM: Instruction struct (#258)
Previously `Instruction` was an enum. That enum is now `InstructionKind` which is a field of the new struct `Instruction`.

This enables adding a `source_offset` field in the future. It would have been very annoying to add a `source_offset` field to each variant of the enum. Now it can be a field of the struct.