Remove unused diesel feature (#338)
I never noticed an issue when switching to CockroachDB, so I suspected that this feature wasn't actually being used. So, I disabled it and there were no issues. We never serialize these things to the DB anymore, and if we do it would be as JSON which we already have.