Add new feature 'derive-jsonschema-on-enums'
I've previously found that deriving JsonSchema on our big enums (of all
modeling commands and all modeling command responses) takes compilation
from 15s to 75s in release mode.

I tried to make all JsonSchema support conditional (via Cargo features)
but realized that was a bad idea because basically every consumer of
this crate needs JsonSchema support.

But, I recently realized that very few consumers need JsonSchema derived
ON THE BIG ENUMS. Only the servers need that derive. The clients (like
cli and kcl-lib) don't need that derive. They still want to derive
JsonSchema on each individual modeling command (which is very fast) but
not on the enum of all commands (which is slow).