Test OpenAPI schema has no breaking changes (#979)
Modified `test_openapi` so that after it generates the current schema and runs the OpenAPI linter, it'll also check for breaking changes.

I'm adding this test because we can no longer break old versions of Zoo Design Studio. We should really never break the OpenAPI schema. If you find you need a breaking change, just make a new endpoint instead.

The new "check for breaking changes" test logic is pretty simple. It does this by:

1. Downloading the latest schema on `main` branch from GitHub
2. Reading this branch's schema from disk (written earlier in `test_openapi`)
3. Run `just breaking-api-changes`, which
4. compares the two schemae via the `openapi-changes` CLI tool. 
5. If there's a breaking change, that tool terminates with exit code 1 (and 0 i.e. OK if there's no changes, or only additive changes).

In my testing, this tool has much better understanding of OpenAPI than `oasdiff`, which the engine currently uses. For example, `oasdiff` thinks that adding a new optional field, or a field with a default value is a breaking change. That's wrong. `openapi-changes` correctly classifies those as additions, not breaking changes.